+++ xml_processor.rb
@@ -9,7 +9,7 @@ class XmlProcessor
     @noop = false
   end
   attr_accessor :rename_agenda
-  
+
   def run(noop: false)
     @noop = noop
 
@@ -17,30 +17,23 @@ class XmlProcessor
     replace_on_resouce_node
   end
 
-  def set_rename(&block)
-    define_singleton_method('rename', &block)
-  end
-
   def resource_names
-    if @resource_names
-      return @resource_names
-    end
+    return @resource_names if @resource_names
 
     doc = Nokogiri::XML(File.read(@filepath))
-    doc.css('resources > image').map{ |image|
+    doc.css('resources > image').map do |image|
       image['name']
-    }
+    end
   end
 
   private
+
   def replace_on_resouce_node
     doc = Nokogiri::XML(File.read(@filepath))
-    doc.css('resources > image').map{ |image|
+    doc.css('resources > image').map do |image|
       image['name'] = rename(image['name'])
-    }
-    if !@noop
-      File.write(@filepath, doc.to_xml(indent: 4, encoding: 'UTF-8'))
     end
+    File.write(@filepath, doc.to_xml(indent: 4, encoding: 'UTF-8')) unless @noop
   end
 
   def replace_on_objects_node
@@ -48,17 +41,15 @@ class XmlProcessor
       resource_names.each do |resource_name|
         # image= と backgroundImage= について置き換えを行うがほかがあるかは未調査
         # userLabel="btn_backgorund">が残ったりする
-        line.gsub!(/([iI]mage=)"(#{resource_name})"/) { 
-          $1 + '"' +  rename(resource_name) + '"'
-        }
+        line.gsub!(/([iI]mage=)"(#{resource_name})"/) do
+          Regexp.last_match(1) + '"' + rename(resource_name) + '"'
+        end
       end
     end.join("\n") + "\n"
-    if !@noop
-      File.write(@filepath, new_file_str)
-    end
+    File.write(@filepath, new_file_str) unless @noop
   end
 
-  def rename(original_name)
-    raise NotImplementedError.new("You must implement #{self.class}##{__method__}")
+  def rename(_original_name)
+    raise NotImplementedError, "You must implement #{self.class}##{__method__}"
   end
 end
